+++ packages/gatsby-plugin-cms/src/node-api/cms/fetchNodes.ts
@@ -37,14 +37,6 @@ fragment PageContentFragment on PageContent {
     name
     props
   }
-  beforeBlocks {
-    name
-    props
-  }
-  afterBlocks {
-    name
-    props
-  }
   extraBlocks {
     name
     blocks {