+++ packages/store-ui/src/typings/index.ts
@@ -1,0 +1,5 @@
+export type {
+  PolymorphicProps,
+  PolymorphicComponentProps,
+  PolymorphicComponentPropsWithRef,
+} from './PolymorphicGenerics'