+++ packages/gatsby-source-vtex/CHANGELOG.md
@@ -3,6 +3,66 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 1.1.7 (2021-09-13)
+
+
+### Features
+
+* **store-ui:** Add Skeleton Atom ([#911](https://github.com/vtex/faststore/issues/911)) ([ef25cde](https://github.com/vtex/faststore/commit/ef25cdea6f6e344b4005ecd2ec0c80d959df35b0))
+
+
+
+
+
+## 1.1.6 (2021-09-09)
+
+**Note:** Version bump only for package @vtex/gatsby-source-vtex
+
+
+
+
+
+## 1.1.5 (2021-09-09)
+
+
+### Features
+
+* **store-sdk:** custom events, tests and docs ([#937](https://github.com/vtex/faststore/issues/937)) ([c268542](https://github.com/vtex/faststore/commit/c2685425b8a2b6c40ce2f79a5a0afc2b2022760d))
+
+
+
+
+
+## 1.1.4 (2021-09-09)
+
+
+### Features
+
+* support lambdas ([#933](https://github.com/vtex/faststore/issues/933)) ([0f14557](https://github.com/vtex/faststore/commit/0f14557f7b98635bc82ede499b12cc28f2f51b6e))
+
+
+
+
+
+## 1.1.3 (2021-09-08)
+
+**Note:** Version bump only for package @vtex/gatsby-source-vtex
+
+
+
+
+
+## 1.1.2 (2021-09-06)
+
+
+### Features
+
+* **store-ui:** adding list atom ([#918](https://github.com/vtex/faststore/issues/918)) ([72c3b2e](https://github.com/vtex/faststore/commit/72c3b2eeac85c784bbee310898b24823c2c18c3a))
+
+
+
+
+
 ## 1.1.1 (2021-09-01)
 
 **Note:** Version bump only for package @vtex/gatsby-source-vtex